pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "code-chat-agent"
version = "0.1.0"
description = "A chat agent that answers questions across many repos and languages without hallucinating"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Andrew", email = "andrew@example.com"},
]
keywords = ["chat", "agent", "code", "search", "repositories"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "pydantic>=2.0",
    "python-dotenv>=1.0",
    "httpx>=0.25",
    "click>=8.1",
    "weaviate-client>=4.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "pytest-cov>=4.1",
    "pytest-asyncio>=0.21",
    "mypy>=1.5",
    "ruff>=0.1",
    "black>=23.9",
]

llm = [
    "openai>=1.0",
    "anthropic>=0.7",
]

search = [
    "tree-sitter>=0.20",
]

[project.scripts]
code-chat = "src.code_chat_agent.cli:cli"

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.ruff]
line-length = 100
target-version = "py310"
select = ["E", "F", "W", "I", "UP", "S"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "--cov=src --cov-report=term-only --cov-report=html"
asyncio_mode = "auto"
